refactor: move DataSourceStatus ErrorInfo C Bindings into common
